Revert SwingMouseAndKeyHandler, apply fixes to JOGLMouseAndKeyHandler
